refactor: remove locf.modules and add assertions in bttf.core;
